pyproject.toml
[project]
name = "rachu-ciach-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.8",
    "factory-boy>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.2",
]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
